requirements.txt
langchain>=0.1.9
langgraph>=0.0.26
pydantic>=2.0.0
langchain-openai>=0.0.2
langchainhub
langchain-community>=0.0.27
langchain-cli
langchain-experimental
langserve[all]
tavily-python
openai
python-dotenv
google-search-results
faiss-cpu
pypdf
tqdm
wikipedia
arxiv
atlassian-python-api
numexpr
beautifulsoup4
lxml
tiktoken
ipydatagrid
bqplot
nbstripout
duckduckgo-search
pre-commit
nbdime
sentence-transformers
torch
transformers
qdrant_client
rank_bm25
pandas
duckdb
networkx